requirements.txt
Flask==1.1.2
kubernetes==21.7.0
PyMySQL==1.0.2
pytz==2021.3
PyYAML==6.0.1
requests==2.27.1
